feat: add support for downloading signature from remote
